ActiveStorage Settings
ActiveStorage is a tool that makes it easy to upload files to cloud storages such as AWS, Google Cloud Storage or Microsoft Azure Storage and attach those files by converting them into Active Record objects. It comes with a local disk-based service for development and testing, and supports file duplication to other services for backups and migrations.

With ActiveStorage, an application can transform image uploads, generate image representations of non-image uploads such as PDF files and videos, and extract metadata from arbitrary files.

For all this ActiveStorage uses two tables in the database called active_storage_blobs and active_storage_attachments. After creating a new application, you have to run the following command:

rails active_storage:install

And completing with the following command to be able to generate the tables through a migration.

rails db:migrate

We declare the storage service that we will use in config/storage.yml, the configuration will depend on the service that we use.

After this you have to indicate which service will be used by configuring Rails.application.config.active_storage.service in the path config/environments/development.rb