MVC generation for songs
The structure of models, views and controllers was generated through the following command:

rails generate scaffold Song name:text artist:text album:text genre:text

And completing with the following command to be able to generate the tables through a migration.

rails db:migrate

Let’s start by attaching a single image and a single audio to our model. Add a single line of code to your model definition:

has_one_attached :audio
has_one_attached :image

In songs_controller.rb that was kindly created for us by a scaffold builder. All we have to do here is whitelist an audio and image parameter inside the song_params idiomatic method and make it look like this:

    def song_params
      params.require(:song).permit(:name, :artist, :album, :genre, :audio, :image)
    end

If you want to attach a file to the existing model somewhere else in the controller code, this is how you do it:

@song.audio.attach(params[:audio])
@song.image.attach(params[:image])

Now for the views: in a generated _form.html.erb add a file_field right above the submit button:

  <div class="field">
    <%= form.label :image %>
    <%= form.file_field :image %>
  </div>

  <div class="field">
    <%= form.label :audio %>
    <%= form.file_field :audio %>
  </div>

And in order to display our image and listen to our song:

<% if @song.image.attached? %>
  <p>
    <strong>Image:</strong>
    <br>
    <%= image_tag @song.image %>
  </p>
<% end %>

<% if @song.audio.attached? %>
  <%= audio_tag url_for(@song.audio), controls: true, preload: "none", id: "music" %>
<% end %>